+++ jexxa-core/src/main/java/io/jexxa/infrastructure/drivingadapter/rest/RESTfulRPCAdapter.java
@@ -1,13 +1,10 @@
 package io.jexxa.infrastructure.drivingadapter.rest;
 
 import static io.jexxa.infrastructure.drivingadapter.rest.RESTfulRPCConvention.createRPCConvention;
+import static io.jexxa.utils.json.gson.GsonConverter.getGsonBuilder;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.ZonedDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
@@ -15,14 +12,10 @@ import java.util.Optional;
 import java.util.Properties;
 
 import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
 import com.google.gson.JsonArray;
-import com.google.gson.JsonDeserializer;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-import com.google.gson.JsonPrimitive;
-import com.google.gson.JsonSerializer;
 import io.javalin.Javalin;
 import io.javalin.core.JavalinConfig;
 import io.javalin.http.Context;
@@ -340,7 +333,7 @@ public class RESTfulRPCAdapter implements IDrivingAdapter
             javalinConfig.addStaticFiles(properties.getProperty(STATIC_FILES_ROOT));
         }
 
-        this.openAPIConvention = new OpenAPIConvention(properties, javalinConfig,getGsonBuilder() );
+        this.openAPIConvention = new OpenAPIConvention(properties, javalinConfig, getGsonBuilder() );
     }
 
     private Server getServer()
@@ -375,42 +368,5 @@ public class RESTfulRPCAdapter implements IDrivingAdapter
         return sslContextFactory;
     }
 
-    private static GsonBuilder getGsonBuilder()
-    {
-        var gsonBuilder = new GsonBuilder();
-        registerTypeAdapter(gsonBuilder);
-        return gsonBuilder;
-    }
 
-    private static void registerTypeAdapter(GsonBuilder gsonBuilder)
-    {
-        gsonBuilder.registerTypeAdapter(LocalDate.class,
-            (JsonDeserializer<LocalDate>) (json, type, jsonDeserializationContext) -> {
-                if (json.isJsonPrimitive())
-                {
-                    return LocalDate.parse(json.getAsJsonPrimitive().getAsString());
-                }
-                return LocalDate.of(json.getAsJsonObject().get("year").getAsInt(), json.getAsJsonObject().get("month").getAsInt(), json.getAsJsonObject().get("day").getAsInt());
-            });
-
-        gsonBuilder.registerTypeAdapter(LocalDate.class,
-                (JsonSerializer<LocalDate>) (src, typeOfSrc, serializationContext) -> new JsonPrimitive(src.toString()));
-
-
-        gsonBuilder.registerTypeAdapter(LocalDateTime.class,
-                (JsonDeserializer<LocalDateTime>) (json, type, jsonDeserializationContext) -> LocalDateTime.parse(json.getAsJsonPrimitive().getAsString()));
-
-        gsonBuilder.registerTypeAdapter(LocalDateTime.class,
-                (JsonSerializer<LocalDateTime>) (src, typeOfSrc, serializationContext) -> new JsonPrimitive(src.toString()));
-
-        gsonBuilder.registerTypeAdapter(ZonedDateTime.class,
-                (JsonDeserializer<ZonedDateTime>) (json, type, jsonDeserializationContext) -> ZonedDateTime.parse(json.getAsJsonPrimitive().getAsString()));
-
-        gsonBuilder.registerTypeAdapter(ZonedDateTime.class,
-                (JsonSerializer<ZonedDateTime>) (src, typeOfSrc, serializationContext) ->
-
-             new JsonPrimitive( DateTimeFormatter.ISO_OFFSET_DATE_TIME.
-                    format(src.withZoneSameInstant(src.getZone())) ));
-
-    }
 }