+++ jexxa-core/src/main/java/io/jexxa/utils/json/JSONManager.java
@@ -1,0 +1,23 @@
+package io.jexxa.utils.json;
+
+import io.jexxa.utils.json.gson.GsonConverter;
+
+public class JSONManager
+{
+    public static JSONConverter jsonConverter = new GsonConverter();
+
+    public static JSONConverter getJSONConverter()
+    {
+        return jsonConverter;
+    }
+
+    public static void setJSONConverter(JSONConverter jsonConverter)
+    {
+        JSONManager.jsonConverter = jsonConverter;
+    }
+
+    private JSONManager()
+    {
+        //Private constructor
+    }
+}